wiretap: add a dump-close API that provides a "needs to be reloaded" indication.
This allows the "needs to be reloaded" indication to be set in the close
process, as is the case for ERF; having a routine that returns the value
of that indication is not useful if it gets seet in the close process,
as the handle for the wtap_dumper is no longer valid after
wtap_dump_close() finishes.

Fixes #17989.

(backported from commit 109b92b5d796f2fb64491fe1976257bc976f1f40)